requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
spotipy>=2.23.0
python-dotenv>=1.0.0
customtkinter>=5.2.0
Pillow>=10.0.0
sounddevice>=0.4.6
vosk>=0.3.45
